requirements.txt
# Core Django
Django==5.2.6
djangorestframework==3.16.1
